feat: 메인 페이지 상단부 레이아웃 컴포넌트화
